refactor so that main.py and langchain_application.py can use the same config.py
