+++ main.py
@@ -4,23 +4,7 @@ import shutil
 from app_modules.overwrites import postprocess
 from app_modules.presets import *
 from clc.langchain_application import LangChainApplication
-
-
-# 修改成自己的配置！！！
-class LangChainCFG:
-    llm_model_name = 'THUDM/chatglm-6b-int4-qe'  # 本地模型文件 or huggingface远程仓库
-    embedding_model_name = 'GanymedeNil/text2vec-large-chinese'  # 检索模型文件 or huggingface远程仓库
-    vector_store_path = './cache'
-    docs_path = './docs'
-    kg_vector_stores = {
-        '中文维基百科': './cache/zh_wikipedia',
-        '大规模金融研报': './cache/financial_research_reports',
-        '初始化': './cache',
-    }  # 可以替换成自己的知识库，如果没有需要设置为None
-    # kg_vector_stores=None
-    patterns = ['模型问答', '知识库问答']  #
-    n_gpus=1
-
+from config import LangChainCFG
 
 config = LangChainCFG()
 application = LangChainApplication(config)
